Change how global.json is located to account for Helix layout.
The DotNetSdkLocator was coded to expect the the normal development
layout of files, the test assembly is deeply nested under /artifacts.
The Helix layout is quite different with test assemblies nested much
shallower. Workspacese need to support both environments and this
can be acheived by walking up the filesystem from the test assembly.